+++ src/components/TreeView/tests/TreeSearch.test.tsx
@@ -1,12 +1,12 @@
-import { render, RenderResult, screen } from "@testing-library/react";
+import { render, screen } from "@testing-library/react";
 import { act, renderHook } from "@testing-library/react-hooks";
 import userEvent from "@testing-library/user-event";
 import React from "react";
 import { Key } from "ts-key-enum";
 
 import "tests/mockReactI18next";
+import { SemanticDomainTreeNode } from "api";
 import * as backend from "backend";
-
 import TreeSearch, {
   insertDecimalPoints,
   testId,
@@ -14,7 +14,6 @@ import TreeSearch, {
   useTreeSearch,
 } from "components/TreeView/TreeSearch";
 import domMap, { mapIds } from "components/TreeView/tests/MockSemanticDomain";
-import { SemanticDomainTreeNode } from "api";
 import { newSemanticDomainTreeNode } from "types/semanticDomain";
 
 // Handles
