+++ src/components/TreeView/tests/TreeViewActions.test.tsx
@@ -1,13 +1,14 @@
+import configureMockStore from "redux-mock-store";
+import thunk from "redux-thunk";
+
+import * as backend from "backend";
 import {
   setDomainLanguageAction,
   traverseTreeAction,
-  TreeActionType,
 } from "components/TreeView/TreeViewActions";
 import { defaultState } from "components/TreeView/TreeViewReducer";
 import { newSemanticDomainTreeNode } from "types/semanticDomain";
-import * as backend from "backend";
-import configureMockStore from "redux-mock-store";
-import thunk from "redux-thunk";
+import { TreeActionType } from "../TreeViewReduxTypes";
 
 jest.mock("backend", () => ({
   getSemanticDomainTreeNode: (id: string, lang: string) =>
